feat: moved all const in the components to constants folder
