+++ src/components/Fields.jsx
@@ -1,7 +1,5 @@
 import clsx from 'clsx'
-
-const formClasses =
-  'block w-full appearance-none rounded-md border border-gray-200 bg-gray-50 px-3 py-2 text-gray-900 placeholder-gray-400 focus:border-blue-500 focus:bg-white focus:outline-none focus:ring-blue-500 sm:text-sm'
+import { FORM_CLASSES } from '@/constants'
 
 function Label ({ id, children }) {
   return (
@@ -24,7 +22,7 @@ export function TextField ({
   return (
     <div className={className}>
       {label && <Label id={id}>{label}</Label>}
-      <input id={id} type={type} {...props} className={formClasses} />
+      <input id={id} type={type} {...props} className={FORM_CLASSES} />
     </div>
   )
 }
@@ -33,7 +31,7 @@ export function SelectField ({ id, label, className = '', ...props }) {
   return (
     <div className={className}>
       {label && <Label id={id}>{label}</Label>}
-      <select id={id} {...props} className={clsx(formClasses, 'pr-8')} />
+      <select id={id} {...props} className={clsx(FORM_CLASSES, 'pr-8')} />
     </div>
   )
 }
