+++ src/components/Filters.jsx
@@ -1,111 +1,15 @@
 import { useState, useEffect } from 'react'
-import { Dropdown, DropdownItem, Card, Title } from '@tremor/react'
-import { FriendsIcon, ManIcon, NoGenderIcon, WomanIcon } from './Icons'
-import { IconBeach, IconHome, IconSchool, IconBuildingSkyscraper, IconSquareHalf, IconAntennaBars5, IconAntennaBars3, IconUserCode, IconFilter, IconAward, IconDirections, IconRocket } from '@tabler/icons-react'
+import { Dropdown, DropdownItem, Card } from '@tremor/react'
+
 import dynamic from 'next/dynamic'
 
 import { SalariesSectionTitle } from './SalariesSectionTitle'
+import { IconFilter } from '@tabler/icons-react'
+import { EXPERIENCE_FILTERS, GENDERS_FILTERS, MODALITY_FILTERS, STUDIES_FILTERS } from '@/constants'
 
 const AnimatedNumbers = dynamic(() => import('react-animated-numbers'), {
   ssr: false
 })
-
-const EXPERIENCE = [
-  {
-    value: 'all',
-    text: 'Cualquier experiencia',
-    icon: () => <IconUserCode stroke={1} />
-  },
-  {
-    value: 'trainee',
-    text: 'Trainee',
-    icon: () => <IconSchool stroke={1} />
-  },
-  {
-    value: 'junior',
-    text: 'Junior',
-    icon: () => <IconAntennaBars3 stroke={1} />
-  },
-  {
-    value: 'senior',
-    text: 'Senior',
-    icon: () => <IconAntennaBars5 stroke={1} />
-  }
-]
-
-const MODALITY = [
-  {
-    value: 'all',
-    text: 'Cualquier modalidad',
-    icon: () => <IconUserCode stroke={1} />
-  },
-  {
-    value: 'remote',
-    text: 'Remoto',
-    icon: () => <IconBeach stroke={1} />
-  },
-  {
-    value: 'mostlyRemote',
-    text: 'Casi remoto',
-    icon: () => <IconHome stroke={1} />
-  },
-  {
-    value: 'hybrid',
-    text: 'Híbrido',
-    icon: () => <IconSquareHalf stroke={1} />
-  },
-  {
-    value: 'office',
-    text: 'Oficina',
-    icon: () => <IconBuildingSkyscraper stroke={1} />
-  }
-]
-
-const GENDERS = [
-  {
-    value: 'all',
-    text: 'Cualquier género',
-    icon: () => <FriendsIcon />
-  },
-  {
-    value: 'man',
-    text: 'Hombre',
-    icon: () => <ManIcon />
-  },
-  {
-    value: 'woman',
-    text: 'Mujer',
-    icon: () => <WomanIcon />
-  },
-  {
-    value: 'noGender',
-    text: 'Sin definir',
-    icon: () => <NoGenderIcon />
-  }
-]
-
-const STUDIES = [
-  {
-    value: 'all',
-    text: 'Cualquier estudio',
-    icon: () => <IconUserCode stroke={1} />
-  },
-  {
-    value: 'self',
-    text: 'Autodidacta',
-    icon: () => <IconDirections stroke={1} />
-  },
-  {
-    value: 'formal',
-    text: 'Universidad o Grado',
-    icon: () => <IconAward stroke={1} />
-  },
-  {
-    value: 'bootcamp',
-    text: 'Bootcamp',
-    icon: () => <IconRocket stroke={1} />
-  }
-]
 
 export function Filters () {
   const [selectedExperience, setSelectedExperience] = useState('all')
@@ -138,7 +42,7 @@ export function Filters () {
               value={selectedGender}
             >
               {
-              GENDERS.map(({ value, text, icon }) => (
+              GENDERS_FILTERS.map(({ value, text, icon }) => (
                 <DropdownItem className='gap-x-4' key={value} value={value} text={text} icon={icon} />
               ))
             }
@@ -153,7 +57,7 @@ export function Filters () {
               value={selectedExperience}
             >
               {
-              EXPERIENCE.map(({ value, text, icon }) => (
+              EXPERIENCE_FILTERS.map(({ value, text, icon }) => (
                 <DropdownItem className='gap-x-4' key={value} value={value} text={text} icon={icon} />
               ))
             }
@@ -168,7 +72,7 @@ export function Filters () {
               value={selectedModality}
             >
               {
-              MODALITY.map(({ value, text, icon }) => (
+              MODALITY_FILTERS.map(({ value, text, icon }) => (
                 <DropdownItem className='gap-x-4' key={value} value={value} text={text} icon={icon} />
               ))
             }
@@ -183,7 +87,7 @@ export function Filters () {
               value={selectedStudy}
             >
               {
-              STUDIES.map(({ value, text, icon }) => (
+              STUDIES_FILTERS.map(({ value, text, icon }) => (
                 <DropdownItem className='gap-x-4' key={value} value={value} text={text} icon={icon} />
               ))
             }
