+++ src/components/Happiness.jsx
@@ -1,26 +1,19 @@
+import { DATA_HAPPINESS } from '@/constants'
 import {
   Title,
   Text,
   CategoryBar
 } from '@tremor/react'
 
-const data = {
-  title: 'Felicidad por su salario',
-  metric: 65,
-  subCategoryPercentageValues: [20, 20, 20, 20, 20],
-  subCategroyColors: ['red', 'orange', 'yellow', 'lime', 'green'],
-  subCategoryTitles: []
-}
-
 export function Happiness () {
   return (
     <>
-      <Title className='text-center'>{data.title}</Title>
+      <Title className='text-center'>{DATA_HAPPINESS.title}</Title>
       <Text className='text-center'>Más es mejor</Text>
       <CategoryBar
-        categoryPercentageValues={data.subCategoryPercentageValues}
-        percentageValue={data.metric}
-        colors={data.subCategroyColors}
+        categoryPercentageValues={DATA_HAPPINESS.subCategoryPercentageValues}
+        percentageValue={DATA_HAPPINESS.metric}
+        colors={DATA_HAPPINESS.subCategroyColors}
         className='mt-4'
       />
     </>
